src/validation/validators/email/email-validation.ts
import { InvalidFielError } from '@/validation/errors/invalid-field-error'
import { FieldValidation } from '@/validation/protocols/field-validation'

export class EmailValidation implements FieldValidation {
	constructor(readonly field: string) {}

	validate(input: object): Error | null {
		const emailRegex =
			/^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|.(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/

		if (emailRegex.test(input[this.field])) return null
		if (!input[this.field]) return null
		return new InvalidFielError()
	}
}
